Move fixtures to a separate file
