+++ flask_discord_interactions/tests/test_ping.py
@@ -1,20 +1,3 @@
-import pytest
-
-from flask import Flask
-
-from flask_discord_interactions import (DiscordInteractions,
-                                        Client)
-
-@pytest.fixture
-def discord():
-    app = Flask(__name__)
-    return DiscordInteractions(app)
-
-@pytest.fixture
-def client(discord):
-    return Client(discord)
-
-
 def test_ping(discord, client):
     @discord.command()
     def ping(ctx, pong: str = "pong"):
